# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dbaff4c29bdd95510990c0e
  title: TestowyProjekt
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /document/:
    post:
      x-skaffolder-id: 5dbb028229bdd95510990db3
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dbb025529bdd95510990d93
      x-skaffolder-resource: Document
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Document
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                AddDate:
                  type: integer
                AnswerDate:
                  type: integer
                AnswerExpireDate:
                  type: integer
                DocumentDate:
                  type: integer
                Type:
                  type: number
    get:
      x-skaffolder-id: 5dbb027729bdd95510990d9f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dbb025529bdd95510990d93
      x-skaffolder-resource: Document
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Document
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /document/{id}:
    delete:
      x-skaffolder-id: 5dbb028229bdd95510990db8
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dbb025529bdd95510990d93
      x-skaffolder-resource: Document
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Document
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dbb028229bdd95510990db0
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dbb025529bdd95510990d93
      x-skaffolder-resource: Document
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Document
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    post:
      x-skaffolder-id: 5dbb028229bdd95510990db2
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dbb025529bdd95510990d93
      x-skaffolder-resource: Document
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Document
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Document"
    
  #-----  User  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1b
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dbaff4c29bdd95510990c22
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1c
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dbaff4c29bdd95510990c21
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  VindicationCase  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Document:
      x-skaffolder-id: 5dbb025529bdd95510990d93
      x-skaffolder-id-db: 5dbaff4c29bdd95510990c0f
      x-skaffolder-id-entity: 5dbb025529bdd95510990d8c
      x-skaffolder-url: /document
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        AddDate:
          type: integer
          x-skaffolder-id-attr: 5dbb025529bdd95510990d8d
          x-skaffolder-type: Date
        AnswerDate:
          type: integer
          x-skaffolder-id-attr: 5dbb025529bdd95510990d90
          x-skaffolder-type: Date
        AnswerExpireDate:
          type: integer
          x-skaffolder-id-attr: 5dbb025529bdd95510990d8f
          x-skaffolder-type: Date
        DocumentDate:
          type: integer
          x-skaffolder-id-attr: 5dbb025529bdd95510990d8e
          x-skaffolder-type: Date
        Type:
          type: number
          x-skaffolder-id-attr: 5dbb025529bdd95510990d91
          x-skaffolder-type: Number
          x-skaffolder-enumeration:
            - Pismo
      x-skaffolder-relations: 
      required:
        - _id
    User:
      x-skaffolder-id: 5dbaff4c29bdd95510990c18
      x-skaffolder-id-db: 5dbaff4c29bdd95510990c0f
      x-skaffolder-id-entity: 5dbaff4c29bdd95510990c10
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c13
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c14
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c12
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c16
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c15
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c11
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...